Fix order in substitution commands.
When parsing, the pattern is expected first, followed by the replacement,
but when the 'raw' string is reconstructed, the order is the other way
around due to the argument order when passing the data from change() via
translate() to to_change().

Re-order the arguments so that the reconstructed 'raw' string will contain
the arguments in the order in which they were actually present in the source
file.