Split v_arg0 into v_arg0_da, and re-sort.
The order of the entries in %v_arg0 should now match that of the entries
in %v_arg1, and that of %v_arg0_da that of %v_arg1_da.

Also add 'fork' to %v_arg1 (it was previously only in %v_arg0).
It arguably belongs only in %v_arg0, since it doesn't take an argument,
but there's no sub arg0 or to_arg0, so it would otherwise never be called.
Also, wantarray, which also doesn't take an argument, is also in %v_arg1.