Split up the test 06 into multiple files.
Test 06.mock.t tested a whole lot of things. This not only made the file
long and difficult to take in at a glance, it also meant that I had to
wrap subs that any test needed, even if they weren't needed or only
contributed to clutter in other tests. (For example, testing low-priority
conjunctions would really need wrapping 'done', but that would have
created tons of clutter in all the other tests.)

As a side effect of tests only wrapping the subs they need, extract_stack
had to be changed to skip a variable number of stack frames at the end;
in particular, to_Terran caused an additional stack frame to appear
that wasn't otherwise there, so it only needed to be accounted for if
the test wrapped that sub.

Split up the test into multiple files (one per feature) and adjust the
number of tests along the way.

The test numbers are still the same as before, so most files start with
a test number greater than 0. This may be changed in the future, but it's
not a high priority since it works the way it is.

New tests should start at 0, though.