Tests for 1-arg functions.
There are still several TODO and SKIP tests in here.
The skipped ones are mostly those that would exit the script (e.g. dump,
exit); these could be tested in translate-only mode, capturing STDERR
to a scalar or something.
The TODO ones are mostly ambiguities in the parser; new commands will have
to be thought up for a few and the number parser will have to become
stricter in order not to parse e.g. wa'DIchtIn as {wa'} + {DIchtIn}.
Also, wait() is in the list of 1-arg commands but should be moved to
the 0-arg commands.